Refactor file names for survey pages and components
